pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vid2sim-scene"
version = "1.0.0"
description = "VID2SIM Stream 03 - Scene Assembly: scene.json authoring + exporters"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "coacd==1.0.10",
    "jsonschema>=4.21",
    "numpy>=1.26",
    "pygltflib>=1.16",
    "pyyaml>=6.0",
    "trimesh>=4.4",
    "mujoco>=3.3.2",
    "pydantic>=2.6",
    "Pillow>=10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]
usd = [
    "usd-core>=24.08",
]
gemini = [
    "google-genai>=0.3",
]
qwen = [
    "openai>=1.40",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/scene/exporters"]
